Introduit un macro langage

+++ ConfigNano.h
@@ -11,19 +11,19 @@
 // -> attention l'intensité n'est pas réglable en IO, protéger avec une 
 //    résistance en fonction de la led utilisée
 
-const int MODE_PWM = 0;
-const int MODE_IO = 1;
+const byte MODE_PWM = 0;
+const byte MODE_IO = 1;
 
 // ON ou OFF ou autre valeur (fading)
-const int LIGHT_FLASH = 255;
-const int LIGHT_ON = PWM_FOR_LED;
-const int LIGHT_FAD2 = PWM_FOR_LED/2;
-const int LIGHT_FAD4 = PWM_FOR_LED/4;
-const int LIGHT_FAD8 = PWM_FOR_LED/8;
-const int LIGHT_OFF = 0;
+const byte LIGHT_FLASH = 255;
+const byte LIGHT_ON = PWM_FOR_LED;
+const byte LIGHT_FAD2 = PWM_FOR_LED/2;
+const byte LIGHT_FAD4 = PWM_FOR_LED/4;
+const byte LIGHT_FAD8 = PWM_FOR_LED/8;
+const byte LIGHT_OFF = 0;
 
 // Configuration du mode de chaque sortie (Arduino Nano)
-const int outputMode[] = {
+const byte outputMode[] = {
             /* D2 */ MODE_IO,
             /* D3 */ MODE_PWM,
             /* D4 */ MODE_IO,
@@ -37,19 +37,19 @@ const int outputMode[] = {
 };
 
 // Nombre maximum d'éclairages dans le tableau précédent
-#define maxLights sizeof(outputMode)/sizeof(int)
+#define maxLights sizeof(outputMode)/sizeof(byte)
 
 // Entrée digital D15 permettant de choisir la séquence 1 ou la séquence 2 (Pull-up)
-const int seqPin = 15;
+const byte seqPin = 15;
 
 // Entrée digital D14 permettant de passer en RUN (1) ou en STOP (0)  (Pull-up)
 const int startPin = 14; 
 
 // Entrées digitals utilisateur (Pull-Up)
-const int inputUserPin1 = 16;
-const int inputUserPin2 = 17;
-const int inputUserPin3 = 18;
-const int inputUserPin4 = 19;
+const byte inputUserPin1 = 16;
+const byte inputUserPin2 = 17;
+const byte inputUserPin3 = 18;
+const byte inputUserPin4 = 19;
 
 // Entrée pour le seed Generator
-const int seedPin = 7;
+const byte seedPin = 7;
